Moved Constant to subclass of Continuous
